requirements.txt
fastapi==0.109.1
uvicorn==0.24.0
pydantic==1.10.13
sse-starlette==1.8.2
langchain==0.1.16
langchain-core==0.1.44
langchain-google-vertexai==1.0.1
transformers==4.38.0
google-cloud-aiplatform==1.48.0
